Major structural changes
- Switched to built in Dimension2D and Point2D system
- Changed to Double precision
- First implementations regarding UI

+++ example/src/main/kotlin/me/etheraengine/example/entity/component/Attack.kt
@@ -1,11 +1,11 @@
 package me.etheraengine.example.entity.component
 
 data class Attack(
-    var damage: Float,
-    var range: Float,
+    var damage: Double,
+    var range: Double,
     var damageDelay: Long,
     var damageTimeRange: Long,
-    var knockback: Float
+    var knockback: Double
 ) {
     var lastAttackTime: Long = 0
     var isAttacking = false
